renamed userTextSpacing
