+++ docs/index.html
@@ -7,5 +7,5 @@
 	</head>
 	<body>
 		<div id="rsg-root"></div>
-	<script type="text/javascript" src="build/bundle.abe825c2.js"></script></body>
+	<script type="text/javascript" src="build/bundle.806986b9.js"></script></body>
 </html>
